src/requirements.txt
gunicorn
flask
requests
pytesseract
transformers
datasets
accelerate
google-generativeai